fix: fixed issues with merging branch and fixed assertion function for group filters
