+++ src/test/java/stepDef/AddUserToGroupSteps.java
@@ -7,24 +7,26 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
-import pages.BulkEditGroupPO;
-import pages.GroupBrowserPO;
-import pages.LoginPO;
+import pages.*;
 
 import java.util.List;
 
 public class AddUserToGroupSteps {
     private final GroupBrowserPO groupBrowserPO;
     private final BulkEditGroupPO bulkEditGroupPO;
+    private final DashboardPO dashboardPO;
+    private final UserManagementPO userManagementPO;
     private final WebDriver driver;
     private final LoginPO loginPO;
     public String user = "thecptn";
-    private final String username = "jz";
-    private final String password = "123";
+    private final String username = "jzhang1297";
+    private final String password = "Jia1997$";
 
     public AddUserToGroupSteps() {
         groupBrowserPO = new GroupBrowserPO();
         bulkEditGroupPO = new BulkEditGroupPO();
+        dashboardPO = new DashboardPO();
+        userManagementPO = new UserManagementPO();
         loginPO = new LoginPO();
         driver = DriverFactory.getDriver();
     }
@@ -32,12 +34,12 @@ public class AddUserToGroupSteps {
     @Given("I am logged in and on the group browser page")
     public void iAmOnTheGroupBrowserPage() {
         driver.navigate().to(loginPO.URL);
-        loginPO.enterUsername(username);
-        loginPO.enterPassword(password);
-        loginPO.clickLoginButton();
-        driver.navigate().to(groupBrowserPO.URL);
+        loginPO.login(username, password);
+        dashboardPO.clickSettingBtn();
+        dashboardPO.clickUsermanagementBtn();
         groupBrowserPO.enterPassword(password);
         groupBrowserPO.clickConfirmButton();
+        userManagementPO.clickGroupsButton();
     }
 
     @When("I click on bulk edit group members")
